TST/REF: split test_common into multiple files
